Split build script into a reader and a writer
The code to load features from YAML is turned into the package entry
point, so that other scripts can import and use that. This is similar to
how BCD is set up.
